// src/interfaces/Position.ts
enum Position {
    topLeft = 'topLeft',
    topCenter = 'topCenter',
    topRight = 'topRight',
    centerLeft = 'centerLeft',
    centerCenter = 'centerCenter',
    centerRight = 'centerRight',
    bottomLeft = 'bottomLeft',
    bottomCenter = 'bottomCenter',
    bottomRight = 'bottomRight',
    any = 'any',
}

export default Position;